requirements.txt
pytest==8.3.5
selene
allure-pytest==2.14.2
selenium==4.31.0
requests==2.32.3
pydantic==2.11.3
dotenv==0.9.9
allure-python-commons~=2.14.2
python-dotenv~=1.1.0